Refactoring code using new authentication module and applying authentication in a specific way
