+++ rekono/authentication/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+
+from authentication.models import Authentication
+
+# Register your models here.
+
+admin.site.register(Authentication)
