+++ rekono/authentication/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class AuthenticationConfig(AppConfig):
+    '''Authentication Django application.'''
+
+    name = 'authentication'
