+++ rekono/authentication/enums.py
@@ -1,0 +1,12 @@
+from django.db import models
+
+
+class AuthenticationType(models.TextChoices):
+    '''Supported authentication types.'''
+
+    BASIC = 'Basic'
+    BEARER = 'Bearer'
+    COOKIE = 'Cookie'
+    DIGEST = 'Digest'
+    JWT = 'JWT'
+    NTLM = 'NTLM'
