+++ rekono/targets/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 
-from targets.models import (Target, TargetAuthentication, TargetPort,
-                            TargetTechnology, TargetVulnerability)
+from targets.models import (Target, TargetPort, TargetTechnology,
+                            TargetVulnerability)
 
 # Register your models here.
 
@@ -9,4 +9,3 @@ admin.site.register(Target)
 admin.site.register(TargetPort)
 admin.site.register(TargetTechnology)
 admin.site.register(TargetVulnerability)
-admin.site.register(TargetAuthentication)
