+++ rekono/targets/serializers.py
@@ -1,12 +1,10 @@
 from typing import Any, Dict
 
-from api.serializers import ProtectedStringValueField
 from django.forms import ValidationError
 from rest_framework import serializers
-from security.input_validation import validate_credential
 
-from targets.models import (Target, TargetAuthentication, TargetPort,
-                            TargetTechnology, TargetVulnerability)
+from targets.models import (Target, TargetPort, TargetTechnology,
+                            TargetVulnerability)
 from targets.utils import get_target_type
 
 
@@ -144,42 +142,5 @@ class TargetTechnologySerializer(serializers.ModelSerializer):
             raise ValidationError({
                 'name': 'This name already exists in this target port',
                 'version': 'This version already exists for this technology in this target port'
-            })
-        return attrs
-
-
-class TargetAuthenticationSerializer(serializers.ModelSerializer):
-
-    credential = ProtectedStringValueField(required=True, allow_null=False)
-
-    class Meta:
-
-        model = TargetAuthentication
-        fields = ('id', 'target_port', 'name', 'credential', 'type')
-
-    def validate(self, attrs: Dict[str, Any]) -> Dict[str, Any]:
-        '''Validate the provided data before use it.
-
-        Args:
-            attrs (Dict[str, Any]): Provided data
-
-        Raises:
-            ValidationError: Raised if provided data is invalid
-
-        Returns:
-            Dict[str, Any]: Data after validation process
-        '''
-        attrs = super().validate(attrs)
-        validate_credential(attrs['credential'])
-        if TargetAuthentication.objects.filter(
-            target_port=attrs['target_port'],
-            name=attrs['name'],
-            credential=attrs['credential'],
-            type=attrs['type']
-        ).exists():
-            raise ValidationError({
-                'name': 'This name already exists in this target port',
-                'credential': 'This credential already exists for this technology in this target port',
-                'type': 'This credential type already exists for this technology in this target port'
             })
         return attrs
