+++ rekono/targets/urls.py
@@ -1,8 +1,7 @@
 from rest_framework.routers import SimpleRouter
 
-from targets.views import (TargetAuthenticationViewSet, TargetPortViewSet,
-                           TargetTechnologyViewSet, TargetViewSet,
-                           TargetVulnerabilityViewSet)
+from targets.views import (TargetPortViewSet, TargetTechnologyViewSet,
+                           TargetViewSet, TargetVulnerabilityViewSet)
 
 # Register your views here.
 
@@ -11,6 +10,5 @@ router.register('targets', TargetViewSet)
 router.register('target-ports', TargetPortViewSet)
 router.register('target-technologies', TargetTechnologyViewSet)
 router.register('target-vulnerabilities', TargetVulnerabilityViewSet)
-router.register('target-authentication', TargetAuthenticationViewSet)
 
 urlpatterns = router.urls
