+++ rekono/testing/api/test_targets.py
@@ -1,5 +1,3 @@
-from targets.enums import TargetAuthenticationType
-from targets.models import TargetAuthentication
 from testing.api.base import RekonoApiTestCase
 
 
@@ -155,54 +153,3 @@ class TargetVulnerabilitiesTest(RekonoApiTestCase):
         # Delete target vulnerability
         self.api_test(self.client.delete, f'{self.endpoint}{self.target_vulnerability.id}/', 204)
         self.api_test(self.client.get, f'{self.endpoint}{self.target_vulnerability.id}/', 404)
-
-
-class TargetAuthenticationTest(RekonoApiTestCase):
-    '''Test cases for Target Authentication entity from Targets module.'''
-
-    def setUp(self) -> None:
-        '''Create initial data before run tests.'''
-        self.endpoint = '/api/target-authentication/'                           # Target authentication API endpoint
-        super().setUp()
-        self.initialize_environment()                                           # Initialize testing environment
-        # Data for testing
-        self.used_data = {
-            'target_port': self.target_port.id,
-            'name': self.target_credential.name,
-            'credential': self.target_credential.credential,
-            'type': self.target_credential.type
-        }
-        self.models = {                                                         # Models to test __str__ method
-            self.target_credential: f'{self.target_port.__str__()} - {self.target_credential.name}'
-        }
-
-    def initialize_environment(self) -> None:
-        '''Initialize environment for testing.'''
-        super().initialize_environment()
-        self.target_credential = TargetAuthentication.objects.create(
-            target_port=self.target_port,
-            name='admin',
-            credential='admin',
-            type=TargetAuthenticationType.BASIC
-        )
-
-    def test_create(self) -> None:
-        '''Test target authentication creation.'''
-        self.used_data['name'] = 'regularuser'
-        expected = self.used_data.copy()
-        expected['credential'] = len(self.used_data['credential']) * '*'
-        # Create new target authentication
-        content = self.api_test(self.client.post, self.endpoint, 201, data=self.used_data, expected=expected)
-        self.api_test(self.client.get, f'{self.endpoint}{content["id"]}/', expected=expected)
-
-    def test_invalid_create(self) -> None:
-        '''Test target authentication creation with invalid data.'''
-        self.api_test(self.client.post, self.endpoint, 400, data=self.used_data)  # Target authentication already exists
-        self.used_data['credential'] = ';reverseshell'
-        self.api_test(self.client.post, self.endpoint, 400, data=self.used_data)   # Invalid credential value
-
-    def test_delete(self) -> None:
-        '''Test target authentication deletion feature.'''
-        # Delete target authentication
-        self.api_test(self.client.delete, f'{self.endpoint}{self.target_credential.id}/', 204)
-        self.api_test(self.client.get, f'{self.endpoint}{self.target_credential.id}/', 404)
