+++ rekono/testing/executions/test_executions_from_findings.py
@@ -9,8 +9,8 @@ from input_types.models import InputType
 from projects.models import Project
 from resources.enums import WordlistType
 from resources.models import Wordlist
-from targets.enums import TargetAuthenticationType, TargetType
-from targets.models import Target, TargetAuthentication, TargetPort
+from targets.enums import TargetType
+from targets.models import Target, TargetPort
 from tasks.enums import Status
 from tasks.models import Task
 from tools.enums import IntensityRank, Stage
@@ -201,37 +201,37 @@ class ExecutionsFromFindingsTest(TestCase):
         executions = get_executions_from_findings(targets, self.tool)
         self.assertEqual(expected, executions)
 
-    def test_with_targets_and_authentication(self) -> None:
-        '''Test get_executions_from_findings feature with targets and authentication.'''
-        # Target port
-        tp_1 = TargetPort.objects.create(target=self.target, port=80)
-        tp_2 = TargetPort.objects.create(target=self.target, port=22)
-        # Target authentication
-        ta_1_1 = TargetAuthentication.objects.create(
-            target_port=tp_1,
-            name='admin',
-            credential='password',
-            type=TargetAuthenticationType.BASIC
-        )
-        ta_1_2 = TargetAuthentication.objects.create(
-            target_port=tp_1,
-            name='jwt',
-            credential='jwt',
-            type=TargetAuthenticationType.BEARER
-        )
-        ta_1_3 = TargetAuthentication.objects.create(
-            target_port=tp_1,
-            name='cookie',
-            credential='sessionid',
-            type=TargetAuthenticationType.COOKIE
-        )
-        # Target list to pass as argument
-        targets = [self.target, tp_1, tp_2, ta_1_1, ta_1_2, ta_1_3]
-        # Expected executions
-        expected = [
-            [self.target, tp_1, tp_2, ta_1_1],
-            [self.target, tp_1, tp_2, ta_1_2],
-            [self.target, tp_1, tp_2, ta_1_3],
-        ]
-        executions = get_executions_from_findings(targets, self.tool)
-        self.assertEqual(expected, executions)
+    # def test_with_targets_and_authentication(self) -> None:
+    #     '''Test get_executions_from_findings feature with targets and authentication.'''
+    #     # Target port
+    #     tp_1 = TargetPort.objects.create(target=self.target, port=80)
+    #     tp_2 = TargetPort.objects.create(target=self.target, port=22)
+    #     # Target authentication
+    #     ta_1_1 = TargetAuthentication.objects.create(
+    #         target_port=tp_1,
+    #         name='admin',
+    #         credential='password',
+    #         type=TargetAuthenticationType.BASIC
+    #     )
+    #     ta_1_2 = TargetAuthentication.objects.create(
+    #         target_port=tp_1,
+    #         name='jwt',
+    #         credential='jwt',
+    #         type=TargetAuthenticationType.BEARER
+    #     )
+    #     ta_1_3 = TargetAuthentication.objects.create(
+    #         target_port=tp_1,
+    #         name='cookie',
+    #         credential='sessionid',
+    #         type=TargetAuthenticationType.COOKIE
+    #     )
+    #     # Target list to pass as argument
+    #     targets = [self.target, tp_1, tp_2, ta_1_1, ta_1_2, ta_1_3]
+    #     # Expected executions
+    #     expected = [
+    #         [self.target, tp_1, tp_2, ta_1_1],
+    #         [self.target, tp_1, tp_2, ta_1_2],
+    #         [self.target, tp_1, tp_2, ta_1_3],
+    #     ]
+    #     executions = get_executions_from_findings(targets, self.tool)
+    #     self.assertEqual(expected, executions)
