+++ rekono/tools/executor/executor.py
@@ -3,8 +3,7 @@ import logging
 from executions import utils
 from executions.models import Execution
 from executions.queue import producer
-from targets.models import (TargetAuthentication, TargetTechnology,
-                            TargetVulnerability)
+from targets.models import TargetTechnology, TargetVulnerability
 from tasks.models import Task
 from tools.executor.callback import tool_callback
 from tools.models import Argument, Intensity
@@ -28,8 +27,6 @@ def execute(task: Task) -> None:
     targets.extend(list(TargetTechnology.objects.filter(target_port__target=task.target).all()))
     # Add target vulnerabilities to task targets
     targets.extend(list(TargetVulnerability.objects.filter(target_port__target=task.target).all()))
-    # Add target authentication to task targets
-    targets.extend(list(TargetAuthentication.objects.filter(target_port__target=task.target).all()))
     # Get the executions required for this job based on targets and tool arguments.
     # A job can need multiple executions. For example, if the user includes more than one Wordlist and
     # the tool is Dirsearch that only accepts one wordlist as argument. Rekono will
