+++ rekono/tools/tools/base_tool.py
@@ -4,19 +4,21 @@ import re
 import shutil
 import subprocess
 import uuid
-from typing import Any, Dict, List, Tuple, Union, cast
+from typing import Any, Dict, List, Optional, Tuple, Union, cast
 
+from authentication.models import Authentication
 from django.db.models import Model
 from django.db.models.fields.related_descriptors import \
     ReverseManyToOneDescriptor
 from django.db.models.query_utils import DeferredAttribute
 from django.utils import timezone
 from executions.models import Execution
-from findings.models import Finding, Vulnerability
+from findings.models import Finding, Port, Vulnerability
 from findings.queue import producer
 from findings.utils import get_unique_filter
 from input_types.base import BaseInput
 from rekono.settings import REPORTS_DIR, TESTING
+from targets.models import TargetPort
 from tasks.enums import Status
 
 from tools.exceptions import ToolExecutionException
@@ -179,6 +181,26 @@ class BaseTool:
                         break
         return found, command
 
+    def get_authentication(
+        self,
+        targets_list: List[BaseInput],
+        findings_list: List[Finding]
+    ) -> Optional[Authentication]:
+        authentication = None
+        for source, port_type in [
+            (findings_list, Port),
+            (targets_list, TargetPort),
+        ]:
+            if len(cast(List[BaseInput], source)) > 0:
+                ports = [p for p in cast(List[BaseInput], source) if isinstance(p, port_type)]
+                if len(ports) == 1:
+                    authentication = Authentication.objects.get(
+                        target_port__target=self.execution.task.target,
+                        target_port__port=cast(Union[Port, TargetPort], ports[0]).port
+                    )
+                break
+        return authentication
+
     def get_arguments(self, targets: List[BaseInput], previous_findings: List[Finding]) -> List[str]:
         '''Get tool arguments for the tool command.
 
@@ -197,6 +219,9 @@ class BaseTool:
             'intensity': self.intensity.argument,                               # Add intensity config to the arguments
             'output': self.path_output if self.file_output_enabled else ''      # Add output config to the arguments
         }
+        authentication = self.get_authentication(targets, previous_findings)    # Search related authentication instance
+        if authentication:                                                      # Authentication exists
+            previous_findings.append(authentication)                            # Add authentication instance
         for argument in self.arguments:                                         # For each tool argument
             found, command = self.process_argument(
                 argument,
